debug

+++ 2017/models.py
@@ -185,7 +185,7 @@ def pretrain(chunk_size, model_path, x_path, y_path, loss_path, num_epochs=50
                 x = x.to(device)
                 x = x.reshape(x.shape[1], x.shape[2])
                 print('X:', x.shape)
-                target = target.to(device).long()
+                target = target.to(device).float()
                 target = target.reshape(target.shape[1], target.shape[2])
                 print('Y:', target.shape)
                 output = l1(x)
